+++ src/components/utilities.js
@@ -1,9 +1,15 @@
+import React from 'react';
 import moment from 'moment';
 import queryString from 'qs';
 import {
   get,
   pickBy,
 } from 'lodash';
+import {
+  FormattedDate,
+  FormattedMessage,
+} from 'react-intl';
+
 import { searchTypes } from '../constants';
 
 export function isBookPublicationType(publicationType) {
@@ -26,6 +32,44 @@ export function isBookPublicationType(publicationType) {
   };
 
   return !!publicationTypeIsBook[publicationType];
+}
+
+export function compareCoveragesToBeSortedInDescOrder(coverageObj1, coverageObj2) {
+  return new Date(coverageObj2.beginCoverage) - new Date(coverageObj1.beginCoverage);
+}
+
+export function formatCoverageFullDate({ beginCoverage, endCoverage }) {
+  const startDate = beginCoverage
+    ? <FormattedDate value={beginCoverage} timeZone="UTC" />
+    : '';
+
+  const endDate = endCoverage
+    ? <FormattedDate value={endCoverage} timeZone="UTC" />
+    : <FormattedMessage id="ui-eholdings.date.present" />;
+
+  if (!startDate) {
+    return endCoverage ? endDate : '';
+  } else {
+    return <>{startDate}{' - '}{endDate}</>;
+  }
+}
+
+export function formatCoverageYear({ beginCoverage, endCoverage }) {
+  const startYear = beginCoverage
+    ? <FormattedDate value={beginCoverage} timeZone="UTC" year="numeric" />
+    : '';
+
+  const endYear = endCoverage
+    ? <FormattedDate value={endCoverage} timeZone="UTC" year="numeric" />
+    : '';
+
+  if (!startYear) {
+    return endCoverage ? endYear : '';
+  } else if ((moment.utc(beginCoverage).format('YYYY') === moment.utc(endCoverage).format('YYYY')) || (!endYear)) {
+    return startYear;
+  } else {
+    return <>{startYear}{' - '}{endYear}</>;
+  }
 }
 
 export function isValidCoverage(coverageObj) {