+++ src/features/usage-consolidation-accordion/usage-consolidation-accordion.js
@@ -16,6 +16,7 @@ import {
 import Toaster from '../../components/toaster';
 import UsageConsolidationFilters from './usage-consolidation-filters';
 import UsageConsolidationContentPackage from './usage-consolidation-content-package';
+import UsageConsolidationContentTitle from './usage-consolidation-content-title';
 import UsageConsolidationContentResource from './usage-consolidation-content-resource';
 import { getUsageConsolidation as getUsageConsolidationAction } from '../../redux/actions';
 import { selectPropFromData } from '../../redux/selectors';
@@ -33,6 +34,7 @@ const propTypes = {
   isOpen: PropTypes.bool,
   onFilterSubmit: PropTypes.func.isRequired,
   onToggle: PropTypes.func.isRequired,
+  publicationType: PropTypes.string,
   recordType: PropTypes.string.isRequired,
   usageConsolidation: ucReduxStateShape.UsageConsolidationReduxStateShape.isRequired,
 };
@@ -47,6 +49,7 @@ const UsageConsolidationAccordion = ({
   onFilterSubmit,
   costPerUseData,
   recordType,
+  publicationType,
 }) => {
   const filtersInitialState = {
     year: moment().year(),
@@ -54,7 +57,6 @@ const UsageConsolidationAccordion = ({
   };
 
   const stripes = useStripes();
-  const [accordionContentRef, setAccordionContentRef] = useState(null);
   const [filterData, setFilterData] = useState(filtersInitialState);
 
   const canViewUsageConsolidation = stripes.hasPerm('ui-eholdings.costperuse.view');
@@ -91,9 +93,43 @@ const UsageConsolidationAccordion = ({
   };
 
   const renderContent = () => {
+    const {
+      isLoaded: isCostPerUseDataLoaded,
+      isFailed: isCostPerUseDataLoadingFailed,
+    } = costPerUseData;
+
+    if (!isCostPerUseDataLoaded && !isCostPerUseDataLoadingFailed) {
+      return null;
+    }
+
+    if (isCostPerUseDataLoadingFailed) {
+      return (
+        <div data-test-usage-consolidation-error>
+          <FormattedMessage id="ui-eholdings.usageConsolidation.summary.error" />
+        </div>
+      );
+    }
+
     if (recordType === entityTypes.PACKAGE) {
       return (
         <UsageConsolidationContentPackage
+          costPerUseData={costPerUseData}
+          year={filterData.year}
+        />
+      );
+    } else if (recordType === entityTypes.TITLE) {
+      return (
+        <UsageConsolidationContentTitle
+          costPerUseData={costPerUseData}
+          year={filterData.year}
+          publicationType={publicationType}
+        />
+      );
+    }
+
+    if (recordType === entityTypes.RESOURCE) {
+      return (
+        <UsageConsolidationContentResource
           costPerUseData={costPerUseData}
           year={filterData.year}
         />
@@ -112,10 +148,6 @@ const UsageConsolidationAccordion = ({
     return null;
   };
 
-  if (accordionContentRef) {
-    accordionContentRef.style.margin = '0';
-  }
-
   if (usageConsolidation.isFailed || !canViewUsageConsolidation) {
     return null;
   }
@@ -129,7 +161,6 @@ const UsageConsolidationAccordion = ({
           open={isOpen}
           onToggle={onToggle}
           headerProps={headerProps}
-          contentRef={(n) => setAccordionContentRef(n)}
         >
           <UsageConsolidationFilters
             onSubmit={handleFiltersSubmit}